rdf_delta_consumer/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
azure-functions==1.21.3
certifi==2024.8.30
h11==0.14.0
html5lib-modern==1.2
httpcore==1.0.6
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
MarkupSafe==3.0.2
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.2.0
rdf-delta@https://github.com/Kurrawong/rdf-delta-python/archive/refs/tags/0.2.4.zip
rdflib==7.1.0
sniffio==1.3.1
typing_extensions==4.12.2
